ml/feature-engineering/requirements.txt
pandas==2.1.3
numpy==1.26.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-Levenshtein==0.23.0
scikit-learn==1.3.2

